Extract MNIST TF model from model/Model.py
