+++ nnlo/util/utils.py
@@ -56,9 +56,14 @@ def load_model(filename=None, model=None, weights_file=None, custom_objects={}):
         with open( filename ) as arch_f:
             json_str = arch_f.readline()
             new_model = model_from_json( json_str, custom_objects=custom_objects) 
-    if model is not None:
+        logging.info(f"Load model from filename")
+    elif model is not None:
         new_model = clone_model(model)
-    if weights_file is not None:
+        logging.info(f"Load model from model")
+    elif weights_file is not None:
         new_model.load_weights( weights_file )
+        logging.info(f"Load model from weights_file")
+    else:
+        logging.error(f"Cannot load model: filename, model and weights_file are None")
     return new_model
